* Finished startup speedup. Playlistbrowser was querying the id for every single artist. * Scores are now represented as Int's over the whole GUI, internally still floats (of course).
svn path=/trunk/kdeextragear-1/amarok/; revision=344274

+++ src/collectiondb.h
@@ -44,8 +44,8 @@ class CollectionDB : public QObject
         QStringList artistAlbumList( bool withUnknown = true, bool withCompilations = true );
 
         bool getMetaBundleForUrl( const QString url, MetaBundle *bundle );
-        float addSongPercentage( const QString url, const int percentage );
-        float getSongPercentage( const QString url );
+        uint addSongPercentage( const QString url, const int percentage );
+        uint getSongPercentage( const QString url );
         void updateDirStats( QString path, const long datetime );
         void removeSongsInDir( QString path );
         bool isDirInCollection( QString path );
