+++ src/playlistitem.cpp
@@ -198,7 +198,7 @@ void PlaylistItem::setText( const MetaBundle &bundle )
     setText( Length,  bundle.prettyLength() );
     setText( Bitrate, bundle.prettyBitrate() );
 
-    const float score = CollectionDB().getSongPercentage( bundle.url().path() );
+    const uint score = CollectionDB().getSongPercentage( bundle.url().path() );
     if ( score )
         setText( Score, QString::number( score ) );
 }
